Move WebhookService into existing rest bundle
Drop the standalone org.openhab.core.io.rest.webhook bundle and place
Webhook/WebhookService directly in org.openhab.core.io.rest, alongside
RESTResource and LocaleService, per review feedback.
