+++ bom/openhab-core/pom.xml
@@ -288,12 +288,6 @@
     </dependency>
     <dependency>
       <groupId>org.openhab.core.bundles</groupId>
-      <artifactId>org.openhab.core.io.rest.webhook</artifactId>
-      <version>${project.version}</version>
-      <scope>compile</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.openhab.core.bundles</groupId>
       <artifactId>org.openhab.core.io.websocket</artifactId>
       <version>${project.version}</version>
       <scope>compile</scope>
