--- bundles/org.openhab.core.io.rest.webhook/pom.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
-
-  <modelVersion>4.0.0</modelVersion>
-
-  <parent>
-    <groupId>org.openhab.core.bundles</groupId>
-    <artifactId>org.openhab.core.reactor.bundles</artifactId>
-    <version>5.2.0-SNAPSHOT</version>
-  </parent>
-
-  <artifactId>org.openhab.core.io.rest.webhook</artifactId>
-
-  <name>openHAB Core :: Bundles :: Webhook Interface</name>
-
-</project>
