+++ bundles/org.openhab.core.io.rest/src/main/java/org/openhab/core/io/rest/WebhookService.java
@@ -10,7 +10,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  */
-package org.openhab.core.io.rest.webhook;
+package org.openhab.core.io.rest;
 
 import java.util.concurrent.CompletableFuture;
