+++ bundles/pom.xml
@@ -87,7 +87,6 @@
     <module>org.openhab.core.io.transport.serial.rxtx</module>
     <module>org.openhab.core.io.transport.serial.rxtx.rfc2217</module>
     <module>org.openhab.core.io.transport.upnp</module>
-    <module>org.openhab.core.io.rest.webhook</module>
     <module>org.openhab.core.io.websocket</module>
     <module>org.openhab.core.io.websocket.audio</module>
     <module>org.openhab.core.io.jetty.certificate</module>
